update to embedded-hal v1.0
Updated the library to use the latest version of embedded-hal. Also
updated the examples for the raspberry pi pico and the arduino uno
respectively. Although the SpiDevice trait was recently added to the
embedded-hal crate which would remove the need to pass an SpiBus and a
chip select pin to the MCP2515 constructor, the rp2040-hal does not yet
support this well. Once the embedded-hal-bus crate is updated to support
the rp2040, the MCP2515 should only take ownership of a single SpiDevice
and not the entire SpiBus.